requirements.txt
# Dependencies
streamlit==1.41.1
arxiv==2.1.3
openai==1.59.3
python-dotenv==1.0.1

# LLM Providers
anthropic>=0.40.0
google-generativeai>=0.8.0

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database (MySQL)
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0  # PyMySQL需要
alembic>=1.13.0

# Async Tasks
celery[redis]>=5.4.0
redis>=5.0.0

# Extra Literature Sources
semanticscholar>=0.8.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
python-multipart>=0.0.6
httpx>=0.27.0

# Logging and Monitoring
structlog>=24.1.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
